backend/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
ollama
pandas
numpy